move getRecordIdsOverlappingFeatures to sqlite/localforage in preparation for updates
